abstracted charts component
