+++ src/components/exoplanets/exoplanets.js
@@ -1,8 +1,8 @@
 "use client";
 import React, { useEffect, useState } from "react";
 import { fetchBatchedExoplanets } from "@/services/fetch-datasets";
-import ExoplanetHistogram from "./exoplanet-histogram";
 import { InspectExoplanets } from "./inspect-exoplanets";
+import ExoplanetsCharts from "./exoplanets-charts";
 
 const Exoplanets = () => {
     const [allExoplanets, setAllExoplanets] = useState([]);
@@ -18,40 +18,11 @@ const Exoplanets = () => {
 
     return (
         <section>
-        <h1 className="text-3xl font-bold mb-4">Exoplanets</h1>
-
-        <div className="mb-4 text-gray-700 dark:text-gray-300">
-          We've detected {allExoplanets.length} exoplanets so far! Explore the distributions of their discovery years and sizes below.
-        </div>
-
-        {/* Histogram: number of discoveries by year */}
-        <div className="mb-6">
-          {/* derive numeric discovery years from the full dataset */}
-          <ExoplanetHistogram
-            data={allExoplanets
-              .map((planet) => {
-                const y = Number(planet.discovery_year);
-                return Number.isFinite(y) ? y : NaN;
-              })
-              .filter((value) => !Number.isNaN(value))}
-            width={900}
-            height={300}
-            bins={Math.min(new Set(allExoplanets.map((planet) => planet.discovery_year).filter(Boolean)).size || 10, 80)}
-            title="Exoplanet Discoveries by Year"
-          />
-        </div>
-
-        <div className="mb-6">
-            <ExoplanetHistogram
-            data={allExoplanets.map(planet => Number(planet.radius_earth_radii)).filter(Number.isFinite)}
-            width={900}
-            height={300}
-            bins={100}
-            title={"Exoplanet Radius (Compared to Earth)"}
-            />
-        </div>
-
-
+            <h1 className="text-3xl font-bold mb-4">Exoplanets</h1>
+            <div className="mb-4 text-gray-700 dark:text-gray-300">
+            We've detected {allExoplanets.length} exoplanets so far! Explore the distributions of their discovery years and sizes below.
+            </div>
+            <ExoplanetsCharts allExoplanets={allExoplanets} />
             <InspectExoplanets allExoplanets={allExoplanets} isLoading={isLoading} />
         </section>
     );
